requirements.txt
absl-py==2.1.0
ace_tools==0.0
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dash==2.11.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-svg==0.0.12
dash-table==5.0.0
debugpy==1.8.14
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==2.2.5
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==21.2.0
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
h5py==3.11.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.4.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.6.2
matplotlib-inline==0.1.7
mistune==3.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.9.0.80
opt_einsum==3.4.0
overrides==7.7.0
packaging==24.2
pandas==1.5.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.1.0
platformdirs==4.3.6
plotly==5.11.0
pluggy==1.5.0
prometheus_client==0.22.1
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyglet==2.0.15
Pygments==2.19.1
pyparsing==3.2.1
pytest==6.2.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.23.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
swiplserver==1.0.2
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch==2.2.2
tornado==6.5.1
tqdm==4.64.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==1.26.20
wandb==0.19.8
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
widgetsnbextension==4.0.14
wrapt==1.14.1